+++ src/sipnet/sipnet.c
@@ -405,12 +405,10 @@ void readParamData(ModelParams **modelParamsPtr, const char *paramFile) {
  * @param out File pointer for output
  */
 void outputHeader(FILE *out) {
-  fprintf(out, "Notes: (PlantWoodC, PlantLeafC, Soil and Litter in g C/m^2; "
-               "Water and Snow in cm; SoilWetness is fraction of WHC;\n");
-  fprintf(out, "year day time plantWoodC plantLeafC woodCreation ");
-  fprintf(out, "soil microbeC coarseRootC fineRootC ");
-  fprintf(out, "litter soilWater soilWetnessFrac snow ");
   fprintf(out,
+          "year day time plantWoodC plantLeafC woodCreation "
+          "soil microbeC coarseRootC fineRootC "
+          "litter soilWater soilWetnessFrac snow "
           "npp nee cumNEE gpp rAboveground rSoil rRoot ra rh rtot "
           "evapotranspiration fluxestranspiration minN n2oFlux nLeachFlux\n");
 }
@@ -1773,9 +1771,13 @@ void setupModel(void) {
 }
 
 // See sipnet.h
-void runModelOutput(FILE *out, OutputItems *outputItems, int printHeader) {
-  if ((out != NULL) && printHeader) {
+void runModelOutput(FILE *out, OutputItems *outputItems) {
+  if (out != NULL) {
     outputHeader(out);
+  }
+
+  if (outputItems != NULL) {
+    writeOutputItemHeaders(outputItems);
   }
 
   setupModel();